--- a/src/AngleSharp.Benchmarks/Program.cs
+++ b/src/AngleSharp.Benchmarks/Program.cs
@@ -9,6 +9,14 @@ namespace AngleSharp.Benchmarks
         static void Main(String[] args)
         {
             Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
+            var parserBenchmark = new ParserBenchmark();
+            parserBenchmark.Setup();
+
+            Console.ReadLine();
+
+            parserBenchmark.AngleSharp();
+
+            Console.ReadLine();
 
             BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
         }