--- a/src/AngleSharp.Benchmarks/Program.cs
+++ b/src/AngleSharp.Benchmarks/Program.cs
@@ -4,19 +4,21 @@ using BenchmarkDotNet.Running;
 
 namespace AngleSharp.Benchmarks
 {
+    using System.IO;
+    using Html.Parser;
+    using Text;
+
     static class Program
     {
         static void Main(String[] args)
         {
             Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
-            var parserBenchmark = new ParserBenchmark();
-            parserBenchmark.Setup();
+            var source = new PrefetchedTextSource(File.ReadAllText($@"..\..\..\temp\amazon.html").AsMemory());
+            var parser = new HtmlParser();
+            // var parser = new HtmlParser(ParserBenchmark.HtmlParserOptions);
+            var document = parser.ParseDocument(source);
+            Console.WriteLine(document.ToHtml());
 
-            Console.ReadLine();
-
-            parserBenchmark.AngleSharp();
-
-            Console.ReadLine();
 
             BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
         }