--- a/src/AngleSharp.Benchmarks/Program.cs
+++ b/src/AngleSharp.Benchmarks/Program.cs
@@ -13,13 +13,11 @@ namespace AngleSharp.Benchmarks
         static void Main(String[] args)
         {
             Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
-            var source = new PrefetchedTextSource(File.ReadAllText($@"..\..\..\temp\amazon.html").AsMemory());
-            var parser = new HtmlParser();
-            // var parser = new HtmlParser(ParserBenchmark.HtmlParserOptions);
-            var document = parser.ParseDocument(source);
-            Console.WriteLine(document.ToHtml());
-
-
+            // var source = new PrefetchedTextSource(File.ReadAllText($@"..\..\..\temp\amazon.html").AsMemory());
+            // var parser = new HtmlParser();
+            // // var parser = new HtmlParser(ParserBenchmark.HtmlParserOptions);
+            // var document = parser.ParseDocument(source);
+            // Console.WriteLine(document.ToHtml());
             BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
         }
     }