--- a/src/AngleSharp/Browser/EncodingMetaHandler.cs
+++ b/src/AngleSharp/Browser/EncodingMetaHandler.cs
@@ -27,7 +27,12 @@ namespace AngleSharp.Browser
             if (encoding != null)
             {
                 var document = element.Owner!;
-                document.Source.CurrentEncoding = encoding;
+                if (document.WritableSource == null)
+                {
+                    throw new InvalidOperationException();
+                }
+
+                document.WritableSource.CurrentEncoding = encoding;
             }
         }
 
@@ -49,7 +54,7 @@ namespace AngleSharp.Browser
                     return TextEncoding.Resolve(charset);
                 }
             }
-            
+
             var shouldParseContent = element.HttpEquivalent.Isi(HeaderNames.ContentType);
             return shouldParseContent ? TextEncoding.Parse(element.Content ?? String.Empty) : null;
         }