… src/AngleSharp.Benchmarks/test.html
… src/AngleSharp/Common/BaseTokenizer.cs
… src/AngleSharp/Common/ObjectExtensions.cs
… src/AngleSharp/Css/Parser/CssStringSourceExtensions.cs
… src/AngleSharp/Dom/IDocument.cs
… src/AngleSharp/Dom/IEntityProvider.cs
… src/AngleSharp/Dom/Internal/Document.cs
… src/AngleSharp/Dom/Url.cs
… src/AngleSharp/Html/Dom/Internal/HtmlDocument.cs
… src/AngleSharp/Html/HtmlEntityProvider.cs
… src/AngleSharp/Html/Parser/HtmlDomBuilder.cs
… src/AngleSharp/Html/Parser/HtmlDomBuilderExtensions.cs
… src/AngleSharp/Html/Parser/HtmlParser.cs
… src/AngleSharp/Html/Parser/HtmlParserOptions.cs
… src/AngleSharp/Html/Parser/HtmlTokenizer.cs
… src/AngleSharp/Html/Parser/IHtmlParser.cs
… src/AngleSharp/Html/Parser/TokenizerExtensions.cs
… src/AngleSharp/Html/Parser/Tokens/HtmlTagToken.cs
… src/AngleSharp/Html/Parser/Tokens/HtmlToken.cs
… src/AngleSharp/Text/CharExtensions.cs
… src/AngleSharp/Text/ITextSource.cs
… src/AngleSharp/Text/PrefetchedTextSource.cs
… src/AngleSharp/Text/StringExtensions.cs
… src/AngleSharp/Text/TextSource.cs