# Types of Changes

## Prerequisites

Please make sure you can check the following two boxes:

- [x] I have read the **CONTRIBUTING** document
- [ ] My code follows the code style of this project

## Contribution Type

What types of changes does your code introduce? Put an `x` in all the boxes that apply:

- [ ] Bug fix (non-breaking change which fixes an issue, please reference the issue id)
- [ ] New feature (non-breaking change which adds functionality, make sure to open an associated issue first)
- [x] Breaking change (fix or feature that would cause existing functionality to change)
- [x] My change requires a change to the documentation
- [ ] I have updated the documentation accordingly
- [ ] I have added tests to cover my changes
- [ ] All new and existing tests passed


## Description

I am aware that this is a huge PR, which breaks binary compatibility and requires more work to be merged. Treat it like a demonstration of concepts

If I recall correctly I've touched several different points here, which could be summarized as:

1) Try to not allocate small temp buffers on heap, or if necessary rent it from array pool
2) `PrefetchedTextSource` (name tbd) implements a new readonly interface which allowed to offload burden of managing buffers and encoding to the end user, which in turn allows pretty impressive ram savings
3) Entity provider - removed temp buffer allocations +  FrozenDictionary for faster access (as promised by MS)
4) New options for tokenizer to reduce unwanted tokens and by extension dom creation (saves a lot of RAM)
5) Several instances of lazy init

There might be more, please let me know if you want to clarify anything

Plans for future: 

It might be possible to expose ReadOnlyMemory<Char> instead of strings for HtmlTokens, so it would be possible to lazily materialize strings from attributes. 

It might be possible to rewrite html token as struct, not sure yet

Anyway there are some benchmarks to spice it up a bit:

Before:

|           Method |      UrlTest |         Mean |         Error |       StdDev |      Gen0 |      Gen1 |     Gen2 |   Allocated |
|----------------- |------------- |-------------:|--------------:|-------------:|----------:|----------:|---------:|------------:|
|       AngleSharp |       amazon |     39.44 us |      8.441 us |     0.463 us |    6.6528 |    0.5493 |        - |    54.42 KB |
| AngleSharpTokens |       amazon |     13.72 us |      1.698 us |     0.093 us |    1.6632 |    0.0305 |        - |    13.61 KB |
|       AngleSharp | en.wikipedia | 19,436.33 us |  3,728.473 us |   204.370 us | 1156.2500 | 1031.2500 |  62.5000 | 10318.87 KB |
| AngleSharpTokens | en.wikipedia |  5,635.95 us |    984.095 us |    53.942 us |  570.3125 |  148.4375 |  85.9375 |  5359.01 KB |
|       AngleSharp |      mail.ru |  4,248.72 us |  1,600.504 us |    87.729 us |  343.7500 |  296.8750 |  62.5000 |  3814.67 KB |
| AngleSharpTokens |      mail.ru |  2,807.85 us |    330.585 us |    18.120 us |  222.6563 |   93.7500 |  46.8750 |  2220.71 KB |
|       AngleSharp |          nbc | 29,163.82 us |  2,678.585 us |   146.822 us | 1812.5000 | 1375.0000 | 468.7500 | 19032.07 KB |
| AngleSharpTokens |          nbc |  7,845.44 us |  1,888.340 us |   103.506 us |  718.7500 |  328.1250 | 187.5000 |  9536.88 KB |
|       AngleSharp |      netflix |    724.38 us |     37.833 us |     2.074 us |  107.4219 |   78.1250 |  42.9688 |    956.9 KB |
| AngleSharpTokens |      netflix |    787.11 us |    171.275 us |     9.388 us |   53.7109 |   15.6250 |   7.8125 |   546.17 KB |
|       AngleSharp |      nytimes | 12,301.40 us |  3,173.483 us |   173.949 us | 1093.7500 | 1046.8750 | 375.0000 | 11055.26 KB |
| AngleSharpTokens |      nytimes |  5,705.90 us |    421.291 us |    23.092 us |  484.3750 |  289.0625 | 156.2500 |  6268.98 KB |
|       AngleSharp |      spiegel | 56,868.89 us | 27,154.183 us | 1,488.412 us | 2333.3333 | 1555.5556 | 444.4444 | 20818.93 KB |
| AngleSharpTokens |      spiegel | 11,315.12 us |    820.433 us |    44.971 us | 1078.1250 |  453.1250 | 171.8750 | 10372.09 KB |
|       AngleSharp |           vk |    117.57 us |     39.319 us |     2.155 us |   12.2070 |    1.9531 |        - |   100.16 KB |
| AngleSharpTokens |           vk |     20.66 us |      3.461 us |     0.190 us |    3.1128 |    0.0916 |        - |    25.43 KB |

After:

|                     Method |      UrlTest |         Mean |        Error |     StdDev |     Gen0 |     Gen1 |     Gen2 |  Allocated |
|--------------------------- |------------- |-------------:|-------------:|-----------:|---------:|---------:|---------:|-----------:|
|       AngleSharpPrefetched |       amazon | 11,277.84 us | 7,531.010 us | 412.800 us | 562.5000 | 453.1250 |        - | 4610.63 KB |
| AngleSharpTokensPrefetched |       amazon |     10.26 us |     0.861 us |   0.047 us |   0.3510 |        - |        - |    2.91 KB |
|       AngleSharpPrefetched | en.wikipedia | 11,179.44 us | 3,514.671 us | 192.651 us | 562.5000 | 453.1250 |        - | 4610.62 KB |
| AngleSharpTokensPrefetched | en.wikipedia |  5,060.19 us | 1,580.268 us |  86.620 us | 164.0625 |  15.6250 |        - | 1399.25 KB |
|       AngleSharpPrefetched |      mail.ru |  3,187.99 us |   155.753 us |   8.537 us | 125.0000 |  50.7813 |   7.8125 | 1199.09 KB |
| AngleSharpTokensPrefetched |      mail.ru |  2,565.50 us |   169.018 us |   9.264 us |  54.6875 |  11.7188 |        - |  461.91 KB |
|       AngleSharpPrefetched |          nbc | 12,039.51 us | 6,572.711 us | 360.272 us | 609.3750 | 531.2500 |  93.7500 | 5767.43 KB |
| AngleSharpTokensPrefetched |          nbc |  6,775.77 us |   800.730 us |  43.891 us | 406.2500 | 289.0625 | 132.8125 | 3707.78 KB |
|       AngleSharpPrefetched |      netflix |    660.57 us |    48.176 us |   2.641 us |  37.1094 |  11.7188 |   3.9063 |  365.66 KB |
| AngleSharpTokensPrefetched |      netflix |    655.16 us |    32.163 us |   1.763 us |  16.6016 |   0.9766 |        - |  139.12 KB |
|       AngleSharpPrefetched |      nytimes |  7,449.52 us | 1,380.397 us |  75.664 us | 390.6250 | 335.9375 |  54.6875 | 3765.77 KB |
| AngleSharpTokensPrefetched |      nytimes |  4,800.10 us |   668.548 us |  36.645 us | 226.5625 | 132.8125 |  62.5000 | 2282.83 KB |
|       AngleSharpPrefetched |      spiegel | 18,073.75 us |   413.158 us |  22.647 us | 750.0000 | 625.0000 |        - | 6679.16 KB |
| AngleSharpTokensPrefetched |      spiegel | 10,175.03 us | 2,220.511 us | 121.714 us | 296.8750 |  62.5000 |        - | 2425.65 KB |
|       AngleSharpPrefetched |           vk |     54.03 us |     3.442 us |   0.189 us |   5.9814 |   0.4883 |        - |   49.02 KB |
| AngleSharpTokensPrefetched |           vk |     21.14 us |     1.992 us |   0.109 us |   0.7935 |        - |        - |    6.65 KB |